chore: add tests and move fn to utils
